+++ src/core/src/math/angle.rs
@@ -145,15 +145,6 @@ where
 #[derive(Clone, Copy)]
 pub struct ArcSec<T: BaseFloat>(pub T);
 
-impl<T> ArcSec<T>
-where
-    T: BaseFloat,
-{
-    fn _truncate(&mut self) {
-        *self = Self((*self).trunc());
-    }
-}
-
 impl<T> From<Rad<T>> for ArcSec<T>
 where
     T: BaseFloat,
@@ -409,7 +400,7 @@ where
         S::max_value().to_angle()
     }
 
-    pub fn to_radians(&self) -> S {
+    pub const fn to_radians(&self) -> S {
         self.rad
     }